StyleItem: add "outputProperties" property
This can be used to let the outside world know of some properties that
might be of use. FOr example, we'll soon use it in menu items to publish
the value of the tabWidth, so that the menu can find the maximum value
and use that one for size computation.